move tests, fix lint
